feat(COR-1954): Initial copy-paste merging
